Let telegraf find the hostname if not provided
In some architectures, the hostname will be unknown when using the role.
Telegraf is able to detect the OS hostname and this should be used in
these cases.

Using services like Amazon Web Services Auto Scaling Group, users cannot
run the role on running instances. It will be better to configure
telegraf in an image (AMI in AWS context) and instances will be started
based on this image.
It means actual instances hostnames will be unknown at configuration
time. The current state of the role will write in the config file the
hostname of the instance used to generate the image, so all future
machines will send points with the same hostname.

This patch still allows to specify an hostname to write in the telegraf
configuration but allso let telegraf find the hostname itself on service
start time.